requirements.txt
numpy==1.21.6
scipy==1.5.2
torch==1.7.1
scikit-image==0.18.1
torchvision==0.8.2
matplotlib==3.3.2
imageio==2.9.0
pillow>=8.3.2

